requirements.txt
requests
youtube-transcript-api
beautifulsoup4
google-generativeai
